ref: moved server module into Server.cpp
