+++ include/ClientStatus.hpp
@@ -6,6 +6,8 @@ enum ClientStatus {
   READY_TO_RESPOND
 };
 
+enum HandleResult { KEEP_CONNECTION, DROP_CONNECTION };
+
 const std::string READ_END = "\r\n";
 
 #endif
